% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_predict.R
\name{auto_predict}
\alias{auto_predict}
\title{Generate Price Predictions on New BMW Data}
\usage{
auto_predict(clean_new_data_df)
}
\arguments{
\item{clean_new_data_df}{Data frame used for price prediction after being passed through auto_clean function.}
}
\value{
Returns price predictions for new BMW data.
}
\description{
Generates price predictions on new BMW data using the final model. Note that the data frame
must be passed through the auto_clean function prior to using the auto_predict function.
}
\examples{
auto_predict(auto_clean(data.frame(model = "5 Series", year = 2018, transmission = "Manual", mileage = 50000, fuelType = "Diesel", tax = 100, mpg = 20, engineSize = 2)))

}
\seealso{
\code{\link{auto_clean}}
}
\author{
Nick Bultman, \email{njbultman74@gmail.com}, April 2021
}
\keyword{predict}
